requirements.txt
Flask
pdfplumber
PyMuPDF
requests
pymongo
python-dotenv
